Add kong e2e test
