+++ kong/tests/conftest.py
@@ -3,6 +3,7 @@ import os
 import pytest
 
 from datadog_checks.dev import docker_run
+from datadog_checks.kong import Kong
 
 from . import common
 
@@ -16,3 +17,8 @@ def dd_environment():
         compose_file=os.path.join(common.HERE, 'compose', 'docker-compose.yml'), endpoints=common.STATUS_URL
     ):
         yield common.instance_1
+
+
+@pytest.fixture
+def check():
+    return Kong(common.CHECK_NAME, {}, {})
